+++ packages/blockchain-utils/README.md
@@ -1,2 +1,2 @@
-# truffle-blockchain-utils
+# @truffle/blockchain-utils
 Utilities for identifying and managing blockchains
