+++ packages/truffle-contract/lib/utils.js
@@ -2,7 +2,7 @@ const debug = require("debug")("contract:utils"); // eslint-disable-line no-unus
 const web3Utils = require("web3-utils");
 const bigNumberify = require("ethers/utils/bignumber").bigNumberify;
 const abi = require("web3-eth-abi");
-const BlockchainUtils = require("truffle-blockchain-utils");
+const BlockchainUtils = require("@truffle/blockchain-utils");
 const reformat = require("./reformat");
 
 const Utils = {
