+++ packages/truffle-core/package.json
@@ -70,7 +70,7 @@
   "devDependencies": {
     "glob": "^7.1.2",
     "memorystream": "^0.3.1",
-    "truffle-blockchain-utils": "^0.0.10"
+    "@truffle/blockchain-utils": "^0.0.10"
   },
   "publishConfig": {
     "access": "public"
