+++ packages/truffle-core/test/ethpm.js
@@ -5,7 +5,7 @@ var glob = require("glob");
 var path = require("path");
 var Contracts = require("truffle-workflow-compile");
 var Package = require("../lib/package.js");
-var Blockchain = require("truffle-blockchain-utils");
+var Blockchain = require("@truffle/blockchain-utils");
 var Ganache = require("ganache-core");
 var Resolver = require("truffle-resolver");
 var Artifactor = require("truffle-artifactor");
